Removed SoundcardCapture in favor of AvCapture. The AvCapture class can take one arbitrary video source and another arbitrary audio source. This implicitly tells the backend that the streams from those sources need to be kept in sync.
To identify the audio and video options the classes AudioSource and
VideoSource are used, which give an id, name, description and associated id
(telling you which audio and video sources belong together). I think that part
of the API is OK, but could be better. So if you have a suggestion let me
know.