Fixed an assert that could happen when the mediaSource is deleted
When using streaming, it could happen that the last reference to the
MediaSource is in another thread. So the objects are destroyed from
another thread. In which case we would delete QObject (ioDevice) in
another thread. That is fixed by calling deleteLater which will ensure
that they are deleted in their own thread.
Note: there was a nother assert that could happen due to a race
condition in the worker thread. That is also fixed with this patch.

Reviewed-by: jbache