gstreamer: Do not call gst_deinit() but do free the device and effect manager objects on backend destruction.
Normally the backend is only destroyed once when the app exists, but calling gst_deinit()
means that the whole gstreamer system becomes unrecoverable. This means that if
you switch backends away from gstreamer and then back again it will not work.

Gstreamer devs recommend that you do not call gst_deinit() at all except in test cases.

Also the manager classes were not deleted properly, so this adds this in.