pulseaudio: More internal rework to how we store a cache of the devices from the server.
This changes again the structure so that all output and capture devices are cached with a specific index.
This index will not change from one update to the next such that we have the necessary information
to issue signals etc. when devices become available or disappear.

We also add some methods that match methods from backend interface such that pulse support can be easily added
to the backends.