globalconfig: Cleanup and export the GlobalConfig object.
This replaces any internal implementation details behind a private object which in turn stores the QSettings.
It is important that clients do not think that QSettings is the be-all and end-all for settings storeage,
hense why I've hidden it away (for refrence, when PulseAudio is used the setting file is no longer looked at all)

All inclusions of the file have been updated. For symmetry the phonon/experiemental GlobalSettings class
has also been exported (and it's header file renamed)