pulsesupport: Implement specific stream moves.
This follows a module-device-manager API change to return the index of the device
rather than an available flag. Availability is implied if the index != PA_INVALID_INDEX

We also now subscribe to and grab information for streams as they appear and extract
the phonon.streamid property that is injected into it. This then allows us to
move the streams between devices when required (namely for the Test button, but
other apps may implement this capability too).

Due to the async nature of pulse, the move request may come before the stream actually
exists. To deal with this, we store the move request in a queue and process it
when new streams appear.

The only problem with this approach is that the protocol does not offer a move
operation that tells module-stream-restore to not save the sink moved to.
For our use, we do not want to save things, so I'll discuss with Lennart as to whether
we can bump the protocol and offer a way of doing this officially, or whether
I need to add another operation to my protocol extension in module-device-manager.