pulsesupport: Keep the streamuuid cache alive for the duration of AudioOutput object.
Many applications (e.g. amarok and others) will reuse the same AudioOutput object.
Unfortunatly the backend itself is closed/reopened which casues a new PulseAudio
sink input to be created with a different index. In order to ensure the device
is set correctly when the stream is recreated we shouldn't remove
the uuid->device map onces we have completed our first move, but rather keep
it alive for the whole duration of the AudioOutput object and ensure the device
is correct each time it's reused.