pulsesupport: Add support for older PA versions sans module-device-manager.
This patch was originally written by Harald Fernengel and tweaked by
Rex Dieter to work with version numbers.

I've shuffled a few things around since Haralds initial patch,
renaming a function to be more appropriate and adding a few more
connection log messages. Also moved where some of the code was
executed, but this doesn't change the semantics of the original
patch.

The lowest version is currently set to 0.9.15 but it should
work with earlier versions too. If some one tests this I'd
be happy to lower it, but not got older versions available
to test right now.