gstreamer: Add some extra mime-types for ogg support.
This is a patch I wrote a while ago and have been carrying
in Mandriva packages. It was originally written to allow
the backend to claim it supports a more types relating to ogg
than those that come directly from gstreamer. This stems from
the fact that several different mime types seem to have become
commonly used for Ogg Vorbis.