pulse: Intercept the objectDescription(Indexes|Properties) calls prior to getting to the backend
At present the xine and gstreamer backends both check for and return results from PA
for these methods. It would be cleaner (and less disruptive on the backends) if
we simply didn't call them when we know PA is active.

This commit does this check. It will be followed by cleanup commits in
both xine and gstreamer backends to remove the unneeded checks/calls that
result from this change.