phonon: Ensure the backend object exists before we use GlobalConfig().
As the backends now explicitly enable PA support on initialisation
we cannot use GlobalConfig() until our backend is initialised.

While this is a rather annoying prerequisit, it doesn't seem to
cause any problem in the kdebase-runtime code which also
uses GlobalConfig().

While this raises some doubt in my mind about how well this
all holds together I think that defaulting to not supporting
PA in the backend allows for safer development of backends
where unexpected PA integration in phonon itself is not triggered
unless the backend specifically says to do so.