pulsesupport: Only process stream moves if the stream exists.
The previous method kept track of the device as dictated by Phonon preferences
and ensured that device was always set. This actually duplicates the work
of PA itself and also has the disadvantage that if PA has some higher priority
routing scheme (e.g. a specific override for a given app) then this would
be overridden by Phonon. This was inconsistant as a non-phonon app
(conforming to the same preferences) would not suffer this override.

This approach will only move the stream if it exists and ignores the move
otherwise.

This has a drawback however. The Test button (which basically calls
AudioOutput::setOutputDevice()) will no longer work. In order
to be able to fix this again, I've introduced a new signal that is
emitted whenever we read data about a stream that states what
device the stream is using. We can listing for this signal
and deal with it appropriately in the AudioOutput class. A
future commit will include this change.