audiooutput: Handle stream moves from lower down the audio stack.
When PulseAudio is not used, Phonon must take care of moving streams around
and adhearing to priority lists etc. but when PA is used, it takes care of that
for us.

In order to keep users informed to the same degree as before, I've
introduced a new DeviceChangeType of 'SoundSystemChange'. I use this
to indicate a change of devices from PA and ultimately disable
the other types of internal device changes and trust PA completely.

This has the advantage that individual streams can be moved independently
of the priority preference.

In order to notify users I use similar/identical phrases to the
current notifications however in the case of a device changing,
it is not possible to tell if it was due to a higher priority
device becoming available, or if a specific stream move was performed.
In practice this is not likely to be a major problem although it
will be important to expose the fact that a specific (per-app)
device override is in effect in e.g. kmix's 'Move Stream' submenu
(this submenu is WIP but will be part of KDE 4.5) otherwise
users will likely wonder how to reset this override and return
to their priority list preferences.