pulse: Rework the connection probing logic.
Completely rework the connection to PA to avoid the use of a QEventLoop.
It starts off by trying to connect via a pa_mainloop and in a blocking fashion.
If that fails, then we simply carry on as if PA doesn't exist.
If it succeeds, it reads in all the information about sinks/sources and then
disconnects before reconnecting via the nicely integrated glib main loop.

By avoiding the use of a QEventLoop, it fixes strange problems in some applications
such as okular.

BUG: 228324