pulse: Add a new PulseStream class
This is intended to act as an object capable of emitting signals when
certain events occur. It is created when an AudioOutput is initialised
and it is then able to connect certain signals to it. This is in theory
very similar to the current setup except that the code in AudioOutput
no longer has to match up the streamUuid before taking action as
the signal is only ever emitted for the matching stream.

This approach keeps things nicely abstracted in AudioOutput and
puts all the work into the Pulse* classes. This means
the _k_deviceChanged slot is more generic and not tied to
the streamUuid used for PA.

This approach will allow for further future integration,
namely per-application volume control.