pulse: Implement support for per-application volume control.
When pulseaudio is used, pass the volume adjustments directly
to PA and bypass the backend. This means that software volume
scaling is typically left to PA which means that the flat volume logic
that is enabled by default in PA can work to the maximum extent and
that PA mixers such as pavucontrol and now kmix will show the
application volume correctly and allow you to adjust it properly
(where 'properly' means the volume slider in the application
also adjusts itself in accordance to what you've set).

Currently there is some strange discrepency between what the volume
sent to PA is and what the %age shown in the application
is. I suspect this is due to the internal implementation of
Stevens' law, but I've not yet had time to look at this properly.