volumeslider: Improve the feedback loop protection.
The feedback loop protection that currently exists only protects
from loops originating from the slider GUI itself, not when
the notification comes in. The latter would only ever loop one time
due to the protection currently in place but it is still better to
avoid it.

Also reset the flag when the change is observed as this allows for
the setVolume() command to operate in an async manner (which is
the case if the backend supports PulseAudio)