xine: Initialise volume to 1.
When PulseAudio is used, it's essential that the internal representation
of the volume remains at the maxium. Xine will use this to set the initial volume.

Thankfully, xine's PulseAudio output does not seem to use per-application volume control
as otherwise this 100% would be passed on an prevent PA from saving
volumes properly.

Currently, however one more problem remains with Xine+PA+volume
in that the standard volume widget does not seem to be hooked up
correctly. I'll look at this problem next.
BUG: 235193