pulse: Fix the reading of stream information on startup.
Due to the new probe+reconnect logic, it is possible to initialise
a 'PulseStream' object and start the backend outputting it
prior to our control connection being established. This means that
we miss the subscription event and never actually associate the
PA stream index with our stream object. Whenever something changes
on the stream at the PA end things would kick in and would work fine,
so this was sometimes hard to see and identify.

This change simply invalidates all streams when we're finished connecting
to PA and then loads up all info from PA to set the indexes correctly.

CCBUG: 235193