pulse: Hopefully solve erronious error message relating to notifications and startup.
Some users have reported that they get a message logged saying the 'audio playback device  does not work'.
The first device name seems to be unavailable (see the double space), but looking through
the code, I cannot see the codepath that allows this to be the case.

So this code should protect against it, but I'm still a little confused
as to why this should be needed.