pulse: Some signal emission protection and some debug tweaks.
This commit ensures that the objectDescriptionChanged signal is only emitted
when PA is active. If we try to connect, succeed, but the backend ultimately
cannot support PA, we fail.

At present the GST backend operates a 'hit and hope' policy
with regards to the pulsesink. This needs to be tightened up.

Also added is extra debug info when the support is enabled
and a new debug method which is not called anywhere but can
be useful when debugging code.

CCBUG: 59052