Added support to Phonon MMF backend for playback of Qt resource files
The backend accesses the resource file path via MediaSource::url().
A small patch to Phonon was required to enable this, because by
default, Phonon passes a QIODevice, rather than the resource file
path, to the backend.

The backend uses this path to create a QResource object, through which
the memory buffer into which the resource file has been read can be
accessed.  This buffer is wrapped in a Symbian 8-bit descriptor and
passed to the OpenDesL() function of the appropriate MMF client
utility API.

Playback only works for certain file formats, as the Symbian MIME type
recognizer does not always work.  For example, playback of an audio
WAV resource file works, while playback of an MP3 resource file does
not.

Task-number: QTBUG-6562
Reviewed-by: Justin McPherson