pulse: Protect the PulseSupport object creation by a mutex.
Under certain circumstances, two Phonon objects may be created
at the same time by different threads. The PulseSupport object
is a singleton and thus only one instance exists. When the object
is created it may take some time to perform it's probe.

For that reason, we must lock a mutex prior to actually creating
the object and check again after obtaining the lock that it was
not created while waiting for the lock. This should ensure
that all clients wait for the object to be fully loaded prior
to using it and that only one instance will be created.