PulseSupport to use qWarning if no GLIB event loop is around
PulseSupport is deactivated if there is no glib event loop around
(decativated by user or Qt compiled without it). To reflect this to the
user it used a regular logMessage. The implications of this can be much
greater though, hence usage of qWarning is more appropriate, to be
independent of the conditional printing of a log message depending on
appropriately set debug level.