Revert "dont define QT_NO_DEBUG for all targets on mingw"
This reverts commit d8661982437e47889bd73e887e169e8f0378ad56.

Beware the part that goes like
What may be a bit unclear here ... copy'n'pastable == changes should be done in findkde4internal or done in such a manner that they are not part of such a copy'n'paste section and merely undo whatever was done in a copy'n'pasteable section.

At any rate a rationale for that change would be very nice to have.

CCMail: patrick.vonreth@gmail.com