Hi, my name is branch and I am an alcoholic.
Testing branch for iodevicestreamer with QNR support. QNR additionally to the QIOD atEnd() has a function isFinished(), to allow proper streaming directly from a QNR we need to take both into account when talking about EOD.
Problem with this is that we would introduce linkage against qtnetwork. That would however be a cmake option...