Shoot the MainMenu && add SubtitleMenu
MainMenu, being ambiguous, is now dead and RootMenu takes its place as
supreme item. Apparently Root and Title are sometimes a bit messy (e.g.
on my Interstella DVD root is grossly depending on the backened, and
they all get it wrong...).

Also add SubtitleMenua as those are sometimes not part of the AudioMenu
(also reflect this in the documentation).
AngleMenu is now last as no one supports that junk anyway.