Make DBus optional even when Qt is built with it
Sometimes it is desirable to build and deploy Phonon without dbus even though
the Qt build has dbus support. To achieve that we have a new cmake option
PHONON_NO_DBUS and a new header providing the setting to the world.
Additionally if QT_NO_DBUS is set, we also set P_N_D, making latter the
ultimate macro to go by.

FEATURE: 268127
FIXED-IN: 4.5.1