add new needData function for abstractstreaming
The new needData(qint64) allows backends to request a given amount
of data. Of course they do not always get what they want, but generally
this allows for much less read requests as data size control moves
from the streamer to the backend/mmlibrary which is a good thing.

Note that needData(qint64) by default implements a call to needData()
to keep hassle for consumers as small as possible. In particular
the way things are right now they do not need to implement
needData(qint64) at all, though it of course makes all sorts of sense.