Create and install a PhononConfig.cmake
The targets are now installed in an "export set", and this export set is
installed. Along with it a PhononConfig.cmake which includes the
installed Config-file.
Also the install locations of the libs and headers are not configurable
anymore, but always in include/ and lib/ under CMAKE_INSTALL_PREFIX.

Now the backends can be changed so that they make use of this file.

Alex