As discussed in Randa, don't install anything outside CMAKE_INSTALL_PREFIX by default
Additionally there is a cmake option
PHONON_INSTALL_QT_EXTENSIONS_INTO_SYSTEM_QT (default FALSE), if set to
TRUE, the mkspecs and the designer plugin will be installed into the
system Qt directories.
If not set to TRUE, they will be installed inside CMAKE_INSTALL_PREFIX.
If this differs from the Qt system directories, print a warning message
for the user, so he knows what's going on.

Alex