Move private pimpl macros into own header
This header is then included by phonondefs_p and medianode_p. The reason
for this is that phonondefs_p is including medianode_p, so naturally
it medianode_p could not include phonondefs_p to gain access to the
macro. The only sensible solution to this is that every pimpl macro is
in a dedicated header, so that both the defs and the node can include
the header. This is the apparently last commit in the pimpl cleanup.