Actually require QZeitgeist 0.8.
Adding the version to the feature log will only show it when no
qzeitgeist is found at all (which is not the case when 0.7 is
installed).

We need to add the version to the macro_optional_find_package so that
qzeitgeist is marked as not found when an older version is found.