Introduce a new type of MediaSource, for AVCapture
This type of media source describes an audio device and a video device
together. It's purpouse is to enable the backends to implement audio
video capture, for example capturing video from a webcam and audio from
a microphone, at the same time.

MediaSourcePrivate now has separate device access lists.

Deprecated an old unused protected constructor that stood in the way.

This commit tries to preserve source / abi compatibility.