validate ticks properly
buffering is a concurrent state (unfortunately a bit mis-represented).
this essentially means that you can be playing and buffering or
paused and buffer, so to approximate perfect validation here we simply
store the previous state, if it was playing and we are now in buffering
the tick *may* still be valid (i.e. you concurrently play and buffer)