remove automoc4 dependency
Starting with CMake 2.8.6 there is an automoc builtin. Whenever using
CMake >= 2.8.6 we therefore activate the builtin and provide a
compatibility macro for automoc4_add_library. When using < 2.8.6 we
continue to look for automoc4 and fail if unable to find it.

This makes the automoc4 dep unnecessary when building with a sufficiently
new CMake version. Horray!