manually set cmake_install_prefix on the icon_install_dir
when doing a cache update we apparently need the full path, which may
be limited to this case, not quite sure.
at any rate our code is vastly different from kdelibs' so a resync is
out of the question for phonon4.

would be good if one of the windows guys were to test this change.

BUG: 294505
FIXED-IN: 4.6.1