use qputenv and force override PA props appname/version/icon_name
libvlc takes absolutely no effort to get a consumer's anything, so all
pvlc applications end up being listed as libvlc with vlc icon, to prevent
that from happening we now force override with our fancy name (as provided
by platform) and additionally assume that the regular qapp applicationname
is equal to the icon name (in all other cases it will be difficult to get
this sorted automatically)

technically we could also set the png binary blob PA_PROP_APPLICATION_ICON
by getting the qapp->windowicon, make it a pixmap and safe that to a
qbuffer that points to a qbytearray (qputenv uses qba), unfortunately
something goes wrong in PA at the time of writing so that the blob does
not get picked up. consequently that feature was left out for now.