do not include FindPhononInternal in cmake config
including FindPhononInternal in the cmake config will expose it to the
entire world, which is not what it is meant for. FindPhononInternal is
nothing more than a centralized piece of cmake logic that is shared between
libphonon and all backends. hence the backends should take care of
including it, which is simply enough as the only need to know where the
buildsystem files are installed.