move phonon version stuff out of phononnamespace
having it inside phonnnamespace makes a rather big and important header
into a .h.in which makes it rather hard to find what exactly is cmake
dependent in there. also kdeversion.h is standalone so we surely can
do the same. furthermore the phonon version magic does not depend nor
contribute to anything in phononnamespace so there really is no point.

phononversion.h.in is now cmake configured
phononnamespace.h is the new static version without configure