import relative dir resolution workaround from CMakePackageConfigHelpers
this prevents the install path resolution from going wrong when symlinks
are involved in /usr/lib* (e.g. the case on Fedora).

the result of not having this workaround is that the install dir will be
resolved as '/' resulting in bogus paths for cmake config users (e.g.
include dir would be set to //include/$soname)

note: this was not actually tested on an affected system, weeh