Revert "clang compat fix (from Qt)"
The original Qt commit (3cba3745, "Phonon: Export ObjectDescriptionModel
to prevent link errors with Clang") was a workaround for
http://llvm.org/bugs/show_bug.cgi?id=8177, which was fixed more than 4
years ago.

Now that we have proper support for clang in Phonon and build it with
-fvisibility=hidden, the workaround started causing problems, as
ObjectDescriptionModel::staticObject and others were not being exported
as expected. A quick test for this is building Qt's phonon capabilities
example.

This reverts commit 845e395ab3aba10dbd57207ad71fbbd6b422f107.

REVIEW: 120664