new cmake option PHONON_INSTALL_QT_COMPAT_HEADERS
Enabling this option will install Qt 4 compatibility headers. The option
is not enabled by default as this is purely compatibility one may choose
to want but is not actually part of the original compatibility nonesense
for kdelibs. Also, since this would conflict if one were to have qtphonon
and phonon installed to the same prefix it best be left to distributions
or developers to decide whether they want or need this additional
compatibility.

As with the kdelibs compat includes, the available set of includes is
only meant to cover the includes available in the $oldthing and will not
be expanded to cover new or experimental headers. Additionally usage
is discouraged for the obvious deprecated nature of things.