rename phonon meta include file in the source and instead rename on-install
having phonon and Phonon in the same source dir screws with
case-insensitive operating systems.

note: this still is a problem if one installs as on the target it would
still be the same setup with both cased files in the same directory.
searching for the actual include lineup that debian was trying to solve
appears as though it actualy never existed, so I am actually not sure
where one should put the header to avoid this. perhaps it'd be best to
simply not offer the compat option on !linux. it does certainly make a
lot less sense on other systems anyway.

CCMAIL: kde-mac@kde.org