add new audiooutputinterface signal for 4.9 to propagate muteness changes
as all other interfaces this is an opt-in interface revision. backends
need to mute into this by building with -DPHONON_BACKEND_VERSION_4_9

muteness always had frontend API but was lacking backend API to allow the
backend to drive muteness changes in a reactive manner. this in particular
will allow us to stop using pulsesupport in pvlc and pgst as they can now
drive muteness which previously was only provided by the pulsesupport
overlay class.

REVIEW: 125350