fix slider to behave more async WRT volume changes
when changing the volume by dragging the volume changes are entirely
ignored for as long as the handle is pressed. i.e. we set the volume
but we do not care whether it worked.
upon releasing the handle the volume is read to make sure we have the
right values (there still can be updates after that to correct that)

when changing the volume from outside sources we get updates but do not
reapply them as actions are disabled. i.e. if one changes the slider via
pulseaudio the slider behaves purely reactive.

when changing the volume by scrolling it pretty much behaves purely
reactive, in that wheel->update are two somewhat more separate things
happening. this does however presently have a problem since volumes are
voltage adjusted by phonon we can in fact end up with volumes higher or
lower than what they should be. a qtimer to simulate something like
pressed when dragging might be handy.