Fix bug 356946 - user avatar in Phonon isn't shrinked
As discribed in the bug report: when users has a avatar it's shown in
sound kcm in "Speaker Placement and Testing", but it's not resized.
(e.g. If it is a 1000x1000 image it takes 1000x1000px on screen.) This
simple patch forces maximum size of `KIconLoader::SizeHuge` what is the
same as default icon size.

Aspect ratio isn't preserved, but please note that in many other places
it's also assumed that avatar is a square.

REVIEW: 126438
BUG: 356946