define a compat qt5_use_modules macro when building for qt5
Summary:
Qt 5.11 dropped the macro, we still rely on it as a single point for
managing qt dependencies. we now define a dummy wrapper for it even when
building qt5 if the macro doesn't already exist

Test Plan:
Needs testing with 5.11 beta. Assuming, I remember things correctly this
should fix libphonon, as well as all backends (so long as they are built
against this libphonon)

Reviewers: asturmlechner

Differential Revision: https://phabricator.kde.org/D12276