revise backend path handling
slightly streamline passing into the factory and also communication to
the backends. the latter specifically is implicitly relying on ECM
variables so as to construct valid paths for the backends prefix
regardless of phonon's prefix. meaning they could theoretically be prefixed
in different directories (practically that currently wouldn't work
because of factorization works though... well... unless the backends get
moved out of the prefix and into $bin as on win32)