remove graphicsview and declarative plugin
these were opt-in exclusive features for qtquick1 and graphicsview. they
were never used and so never became default. throw them away

(this is technically BIC when the option was enabled. which, because the
option is off by default, should not affect anyone)

there's still a factorization left-over reference to the
VideoGraphicsObject as that was part of the standard factorization enum
and is not safe to remove.