remove dbus support from audiooutput
the dbus interface was used to coordinate with the phonon KCM but has
actually not been used for at least 5 years and is largely no longer
relevant. device control is in the hands of pulseaudio, so if you
want to control outputs you would do it via PA, not dbus. because device
control is in the hands of PA the dbus api working is a bit hit and miss
depending on what PA thinks it needs to do with a given device.
the second part of the dbus api was for live-reloading of the backend
which hasn't been used, supported or working in probably even longer.

the entire dbus magic was internal and not represented in the API