Future-proof build fix for clang > 16
Clang is intending to remove the `-Wno-enum-constexpr-conversion` flag
in the future, because it can invoke undefined behavior.

To avoid the "integer value NNNN is outside the valid range of values
[0, 7] for the enumeration type 'ObjectDescriptionType'" warnings,
explicitly specify that the `ObjectDescriptionType` enum uses `unsigned`
as the underlying type.

This also allows to remove the CMake patch that added the
`-Wno-enum-constexpr-conversion` flag.